requirements.txt
pydantic==2.5.1
pyyaml==6.0.1
loguru==0.7.2
chainlit==0.7.603
pyautogen==0.2.0b5
PyGithub==2.1.1
GitPython==3.1.40